+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix gender value mix character cases.
 - Fix `fetch_text()` by adding retry ability with exponential backoff. [#85]
+- Fix incorrect name parsing when a nickname or a middle name was specified. [#84]
 
 ## [[1.5.0]] - 2019-03-19
 
@@ -113,4 +114,5 @@ This first version allows a user to retrieve traffic fatality repports for a cer
 [#61]: https://github.com/scrapd/scrapd/pull/61
 [#64]: https://github.com/scrapd/scrapd/pull/64
 [#66]: https://github.com/scrapd/scrapd/pull/66
+[#84]: https://github.com/scrapd/scrapd/pull/84
 [#85]: https://github.com/scrapd/scrapd/pull/85