+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[1.5.1]] - 2019-04-25
+
 ### Fixed
 
 - Fix gender value mix character cases.
@@ -96,6 +98,7 @@ This first version allows a user to retrieve traffic fatality repports for a cer
 [1.4.1]: https://github.com/scrapd/scrapd/releases/1.4.1
 [1.4.2]: https://github.com/scrapd/scrapd/releases/1.4.2
 [1.5.0]: https://github.com/scrapd/scrapd/releases/1.5.0
+[1.5.1]: https://github.com/scrapd/scrapd/releases/1.5.1
 
 [//]: # (Issue/PR links)
 [#13]: https://github.com/scrapd/scrapd/issues/13